requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
supabase
httpx
celery
redis
